+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# dev
+
 # callr 3.1.0
 
 * New `rscript()` function and `rscript_process` class to execute